requirements.txt
asgiref==3.8.1
attrs==24.3.0
cffi==1.17.1
cryptography==44.0.2
Django==5.1.4
django-cors-headers==4.6.0
django-dbbackup==4.2.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
